Moved phone-validation.js from src folder into src/utils
